analyzeDataFrames.py was edited to analyzeData.py. Now, extractData.py will do the extraction of the data and sum it up.
