+++ .gitignore
@@ -1,1 +1,2 @@
 *.swp
+__pycache__
