+++ extractData.py
@@ -604,7 +604,7 @@ class Chainbuild(Extract):
         return pd.Series(nMolecules,index=range(len(nMolecules)))
     def ExtractBoxLengths(self,inputFileName,dimLetter):
         path = self.path
-        sigma = self.sigma_ff*10 #Angstrom
+        sigma = self.sigma_ff #Angstrom
         solidFileName = self.solidFileName
         boxLengths = []
         print('Warning: Lenghts are not calculated by Chainbuild during simulation.')
@@ -627,7 +627,7 @@ class Chainbuild(Extract):
         return pd.Series(boxLengths,index=range(len(boxLengths)))*sigma #Angstrom
     def ExtractVolumes(self,inputFileName):
         path = self.path
-        sigma = self.sigma_ff*10 #Angstrom
+        sigma = self.sigma_ff #Angstrom
         solidFileName = self.solidFileName
         volume = []
         print('Warning: Volume is not calculated by Chainbuild during simulation.')
@@ -668,7 +668,7 @@ class Chainbuild(Extract):
         return pd.Series(temperature,index=range(len(temperature))) #K
     def ExtractDensities(self,inputFileName,logFileName):
         path = self.path
-        sigma = self.sigma_ff*10 #Angstrom
+        sigma = self.sigma_ff #Angstrom
         density = []
         with open(path+inputFileName,'r') as inputFile: fileLines = inputFile.readlines()
         for line in range(len(fileLines)):
@@ -689,6 +689,71 @@ class Chainbuild(Extract):
                         print('Warning: Density was not found from Chainbuild'); density.append(np.nan)
                 break
         return pd.Series(density,index=range(len(density)))/sigma**3 #Angstrom^-3
+class SummarizeDataFrames():
+    def __init__(self,dataFilesPath,groups,sort,joinDataFrames,countFrom):
+        self.dataFilesPath = dataFilesPath
+        self.groups = groups.split()
+        self.joinDataFrames = joinDataFrames
+        self.countDataFrom = countFrom
+        self.sortDataFramesBy = sort
+        self.summedDataFrame, self.summedDataFrames, self.groupedDataFrames = {}, {}, {}
+    def SearchDataFiles(self):
+        dataFilesPath = self.dataFilesPath
+        groups = self.groups
+        countFrom = self.countDataFrom
+        dataFileNames = os.listdir(dataFilesPath)
+        groupedDataFrames = {i:[] for i in groups}
+        for dataFile in dataFileNames:
+            dataFrame = pd.read_csv(dataFilesPath+dataFile,sep='\t')
+            dfFirstRow = dataFrame[:1]
+            dataFrame = pd.concat([dfFirstRow,dataFrame[countFrom:]])
+            for i in groups:
+                if re.search(i,dataFile): groupedDataFrames[i].append(dataFrame)
+        self.groupedDataFrames = groupedDataFrames
+    def JoinDataFrames(self):
+        sort = self.sortDataFramesBy
+        dataFrames = self.summedDataFrames
+        concatDataFrame = pd.concat(dataFrames).reset_index()
+        concatDataFrame.drop(columns={'level_1'},inplace=True)
+        concatDataFrame.rename(columns={'level_0':'Group'},inplace=True)
+        for key in concatDataFrame.columns:
+            findSortValue = re.search(f'^{sort}\[.+',key)
+            if findSortValue: concatDataFrame.sort_values(findSortValue.group(),ignore_index=True,inplace=True)
+        print('Data Frame:\n')
+        print(concatDataFrame)
+        self.summedDataFrame = concatDataFrame
+    def SetSummedDataFrames(self):
+        groups = self.groups
+        summedDataFrames = self.summedDataFrames
+        for i in groups: summedDataFrames[i] = pd.DataFrame(columns=[])
+        self.summedDataFrames = summedDataFrames
+    def CreateSummedDataFrames(self):
+        sort = self.sortDataFramesBy
+        groups = self.groups
+        groupedDataFrames = self.groupedDataFrames
+        summedDataFrames = self.summedDataFrames
+        for i in groups:
+            for var in groupedDataFrames[i][0].columns:
+                summedDataFrames[i][var] = pd.Series([groupedDataFrames[i][j][var].mean() for j in range(len(groupedDataFrames[i]))])
+                summedDataFrames[i][f'delta{var}'] = pd.Series([groupedDataFrames[i][j][var].std() for j in range(len(groupedDataFrames[i]))])
+            summedDataFrames[i].sort_values(sort,ignore_index=True,inplace=True)
+        self.summedDataFrames = summedDataFrames
+    def Extract(self):
+        self.SearchDataFiles()
+        self.SetSummedDataFrames()
+        self.CreateSummedDataFrames()
+        dataFrames = self.summedDataFrames
+        groups = self.groups
+        for i in groups:
+            print('\n'+i)
+            print(dataFrames[i])
+def SumUpDataFrames(dataFilesPath,groups,sortDataFrames,joinDataFrames,countDataFrom):
+    data = SummarizeDataFrames(dataFilesPath,groups,sortDataFrames,joinDataFrames,countDataFrom)
+    data.Extract()
+    if joinDataFrames:
+        data.JoinDataFrames()
+        return data.summedDataFrame
+    else: return data.summedDataFrames
 def Help():
     print('\nDescription:')
     print('\tThis script extracts the output data generated by RASPA or Chainbuild, and saves them as data frames.')
@@ -705,7 +770,8 @@ def Help():
     print('\t\t-p or -P: Print input parameters.')
     print('\t\t-i or -I: Indicate the path where the data file is found. By default: Output/System_0')
     print('\t\t-o or -O: Indicate the "path/fileName" where the extracted data will be printed.')
-    print('\t\t\tBy default, no file is created.')
+    print('\t\t\tIf you have several files in the input path given, this script will extract them all and enumerate them as i_fileName, where i is an integer number.')
+    print('\t\t\tBy default (if you don\'t set output path), no file is created.')
     print('\t\t-d or -D: List of dimensions to extract the box-lengths. By deafault, the three dimensions.')
     print('\t\t-v or -V: Indicate the list of variables that will be extracted from the RASPA\'s data files (per cycle). By default: Rho and P.')
     print('\tFlags allowed only for RASPA:')
@@ -735,12 +801,28 @@ def Help():
     print('\t\tRho: Density in A^-3.')
     print('Example:')
     print('\tThe following command would extract the volume and internal energy of a methane-benzene binary mixture from Output/System_1 from RASPA,')
-    print('\tas well as the pressure (in atm) and the box-length in the x direction.')
-    print('\tThe extracted data would be saved in the file excessProps.dat')
-    print('\tpython3 extractData.py -m Raspa -I Outputs/System_1/ -o ./excessProps.dat -t prod -v U V P L -U atm -d x -p -s P')
+    print('\tas well as the pressures (in atm) and the box-length in the x axis.')
+    print('\tThe extracted data would be saved in the directory grouped_data, and in the files i_excessProps.dat, where i is a number given by this script.')
+    print('\tpython3 extractData.py -m Raspa -I Outputs/System_1/ -o grouped_data/excessProps.dat -t prod -v U V P L -U atm -d x -p -s P')
     print('\tThe following command would extract the volume, number of molecules and fluid-fluid internal energy of nitrogen both from Chainbuild.')
-    print('\tThe extracted data would be saved in the file outData.dat, and sorted according to the number of molecules.')
-    print('\tpython3 extractData.py -m Chainbuild -I ./ -o ./outData.dat -v V N Uff -p -s N')
+    print('\tThe extracted data would be saved in the directory grouped_data and as files outData.dat, and sorted according to the number of molecules.')
+    print('\tpython3 extractData.py -m Chainbuild -I ./ -o ./grouped_data/outData.dat -v V N Uff -p -s N')
+    print('Note: The file name usually works as the group name as each file will be numbered according to the fileName given.')
+    print('\n')
+    print('Once you grouped your data into a single directory, you can use an independent script to sum up the given data and do your own analysis.')
+    print('1.- In your script, import this file: import extractData as ed')
+    print('2.- Give the following information:')
+    print('    \t- string: Path where the data files are located. Let\'s call it "dataFilesPath".')
+    print('    \t- string: List of the groups of the extracted data (separated by spaces). Call it "groups".')
+    print('    \t- string: The variable that you\'ll use to sort your summed data. Call it "sortDataFrames".')
+    print('    \t- string: Indicate from what points you want to do your analysis (to avoid biased data). You can check the extracted figures to see from what points')
+    print('    \t     your system has reached equilibrium. Call it "countDataFrom".')
+    print('    \t- boolean: Indicate if you want to create a single data frame with your grouped data or separate data frames (one per group). Call it "joinDataFrames"')
+    print('3.- Create a dictionary that will contain your summed data: dataFrames = ed.SumUpDataFrames(dataFilesPath, groups, sortDataFrames, countDataFrom, joinDataFrames)')
+    print('    If you don\'t join your data frames, each dataFrames\'s key will be a group (fileName), and inside each group, you will have your summed data frame.')
+    print('    If you join them all, there will be a single data frame with an extra column called "Group".')
+    print('Note: your data frames were created using the library pandas, so you can take advantage of the library to analyze the data.')
+    print('The file analyzeData.py is an example of a script that would do an analysis.')
     exit(1)
 ##########################################################################################################
 if __name__ == '__main__':
